routes: migrate sessions core domain (~10 routes)
Move to src/server/routes/sessions.ts:
- /api/search GET
- /api/sessions GET (list with archived BFS enrichment)
- /api/sessions/:id GET (verbose live + archived branch)
- /api/sessions/:id DELETE
- /api/sessions/:id PATCH (title, colorIndex, projectId, preview, role,
  assistantType, accessory, delegateOf, teamLeadSessionId, archived)
- /api/sessions/:id/wait POST (chunked heartbeat)
- /api/sessions/:id/activate-skill POST
- /api/sessions/:id/tool-grant-request POST

POST /api/sessions and POST /api/sessions/:archivedId/continue stay in
server.ts — they have heavy session-creation / lossless-resume coupling
worth a dedicated follow-up.

Net: server.ts -533 lines (2833 → 2300). No behaviour change.
