+++ src/server/server.ts
@@ -16,7 +16,6 @@ import { SessionManager } from "./agent/session-manager.js";
 import { RateLimiter } from "./auth/rate-limit.js";
 import { validateToken } from "./auth/token.js";
 import { handleWebSocketConnection } from "./ws/handler.js";
-import { getSlashSkill, buildSlashSkillPrompt } from "./skills/slash-skills.js";
 import { TeamManager } from "./agent/team-manager.js";
 import { shouldCreateWorktree } from "./agent/worktree-decision.js";
 import { RoleStore } from "./agent/role-store.js";
@@ -27,7 +26,6 @@ import { initPromptDirs } from "./agent/system-prompt.js";
 import { recordElapsed } from "./agent/profiling.js";
 
 import { initSkillSidecarDir } from "./skills/skill-sidecar.js";
-import { buildActivationHeader } from "./skills/skill-manifest.js";
 import { TaskManager } from "./agent/task-manager.js";
 import { TaskStore } from "./agent/task-store.js";
 import { BgProcessManager } from "./agent/bg-process-manager.js";
@@ -1007,7 +1005,7 @@ async function handleApiRoute(
 	res: http.ServerResponse,
 	sessionManager: SessionManager,
 	config: GatewayConfig,
-	colorStore: ColorStore,
+	_colorStore: ColorStore,
 	prStatusStore: PrStatusStore,
 	teamManager: TeamManager,
 	roleManager: RoleManager,
@@ -1493,323 +1491,6 @@ async function handleApiRoute(
 			json({ ok: true });
 			return;
 		}
-	}
-
-	// GET /api/search
-	if (url.pathname === "/api/search" && req.method === "GET") {
-		const q = url.searchParams.get("q");
-		if (!q) {
-			json({ error: "Missing query parameter 'q'" }, 400);
-			return;
-		}
-		const limit = Math.min(Math.max(1, parseInt(url.searchParams.get("limit") || "20", 10) || 20), 100);
-		const offset = Math.max(0, parseInt(url.searchParams.get("offset") || "0", 10) || 0);
-		const typeParam = url.searchParams.get("type") || "all";
-		const validTypes = new Set(["all", "goals", "sessions", "messages", "staff"]);
-		const type = validTypes.has(typeParam) ? typeParam as "all" | "goals" | "sessions" | "messages" | "staff" : "all";
-		try {
-			const projectId = url.searchParams.get("projectId") || undefined;
-			const projectNames = new Map(projectRegistry.list().map(p => [p.id, p.name]));
-			const results = await projectContextManager.searchAll(q, { type, limit, offset, projectId, projectNames });
-			json(results);
-		} catch (err) {
-			json({ error: `Search failed: ${err}` }, 500);
-		}
-		return;
-	}
-
-	// BFS helper: walk delegateOf, teamLeadSessionId, teamGoalId, and goalId chains
-	// from seed IDs through an archived session pool.
-	function bfsEnrichArchived(seedIds: string[], allArchived: any[]): any[] {
-		const result: any[] = [];
-		const seen = new Set<string>();
-		const queue = [...seedIds];
-		while (queue.length > 0) {
-			const parentId = queue.shift()!;
-			for (const s of allArchived) {
-				if (!seen.has(s.id) && (
-					s.delegateOf === parentId ||
-					s.teamLeadSessionId === parentId ||
-					s.teamGoalId === parentId ||
-					s.goalId === parentId
-				)) {
-					seen.add(s.id);
-					result.push(s);
-					queue.push(s.id);
-				}
-			}
-		}
-		return result;
-	}
-
-	// GET /api/sessions
-	if (url.pathname === "/api/sessions" && req.method === "GET") {
-		const currentGen = projectContextManager.getSessionGeneration();
-		const sinceParam = url.searchParams.get("since");
-		if (sinceParam !== null) {
-			const since = parseInt(sinceParam, 10);
-			if (!isNaN(since) && since === currentGen) {
-				json({ generation: currentGen, changed: false });
-				return;
-			}
-		}
-		const filterProjectId = url.searchParams.get("projectId") || undefined;
-		const registeredProjectIds = new Set(projectRegistry.list().map(p => p.id));
-		let sessions = sessionManager.listSessions().map((s) => ({
-			...s,
-			colorIndex: colorStore.get(s.id),
-		})).filter(s => !s.projectId || registeredProjectIds.has(s.projectId));
-		if (filterProjectId) {
-			sessions = sessions.filter(s => s.projectId === filterProjectId);
-		}
-		// Support ?include=archived to return archived sessions too
-		if (url.searchParams.get("include") === "archived") {
-			// Collect archived sessions across all project contexts
-			const allArchived: typeof sessions = [];
-			for (const ctx of projectContextManager.all()) {
-				const store = ctx.sessionStore;
-				for (const s of store.getArchived()) {
-					allArchived.push({ ...s, colorIndex: colorStore.get(s.id), status: "archived" } as any);
-				}
-			}
-			// Sort by archivedAt descending
-			allArchived.sort((a: any, b: any) => ((b as any).archivedAt ?? 0) - ((a as any).archivedAt ?? 0));
-			// Apply projectId filter if present
-			const filteredArchived = filterProjectId
-				? allArchived.filter((s: any) => s.projectId === filterProjectId)
-				: allArchived;
-
-			// Collect ALL archived sessions for BFS enrichment (not just delegates)
-			const allArchivedForBfs: typeof sessions = [];
-			for (const ctx of projectContextManager.all()) {
-				for (const s of ctx.sessionStore.getArchived()) {
-					allArchivedForBfs.push({ ...s, colorIndex: colorStore.get(s.id), archived: true } as any);
-				}
-			}
-			// Build live goal IDs for BFS seeding
-			const liveGoalIds: string[] = [];
-			for (const ctx of projectContextManager.all()) {
-				for (const g of ctx.goalStore.getLive()) {
-					if (!g.archived) liveGoalIds.push(g.id);
-				}
-			}
-
-			const limitParam = url.searchParams.get("limit");
-			const afterParam = url.searchParams.get("after");
-			if (limitParam) {
-				// Paginated archived sessions
-				const limit = Math.min(Math.max(1, parseInt(limitParam, 10) || 50), 200);
-				const afterCursor = afterParam ? parseInt(afterParam, 10) : undefined;
-				let page = filteredArchived;
-				if (afterCursor !== undefined) {
-					page = page.filter((s: any) => ((s as any).archivedAt ?? 0) < afterCursor);
-				}
-				const total = filteredArchived.length;
-				const hasMore = page.length > limit;
-				const sliced = page.slice(0, limit);
-				const nextCursor = sliced.length > 0 ? (sliced[sliced.length - 1] as any).archivedAt : undefined;
-
-				// BFS: collect archived children reachable from live sessions and goals
-				const liveIdSet = new Set(sessions.map(s => s.id));
-				const archivedDelegatesOfLive = bfsEnrichArchived([...liveIdSet, ...liveGoalIds], allArchivedForBfs);
-
-				json({ generation: currentGen, sessions: [...sessions, ...sliced], total, hasMore, nextCursor, archivedDelegates: archivedDelegatesOfLive });
-			} else {
-				// BFS: collect archived children reachable from live sessions and goals
-				const liveIdSet = new Set(sessions.map(s => s.id));
-				const archivedDelegatesOfLive = bfsEnrichArchived([...liveIdSet, ...liveGoalIds], allArchivedForBfs);
-
-				// Backward compatible: return all archived sessions
-				json({ generation: currentGen, sessions: [...sessions, ...filteredArchived], archivedDelegates: archivedDelegatesOfLive });
-			}
-		} else {
-			// Always include archived children of live sessions/goals so the sidebar
-			// can render chevrons/nesting without a separate fetch.
-			const liveIdSet = new Set(sessions.map(s => s.id));
-			const allArchivedForBfsNonPaginated: typeof sessions = [];
-			for (const ctx of projectContextManager.all()) {
-				for (const s of ctx.sessionStore.getArchived()) {
-					allArchivedForBfsNonPaginated.push({ ...s, colorIndex: colorStore.get(s.id), archived: true } as any);
-				}
-			}
-			// Build live goal IDs for BFS seeding
-			const liveGoalIdsNonPaginated: string[] = [];
-			for (const ctx of projectContextManager.all()) {
-				for (const g of ctx.goalStore.getLive()) {
-					if (!g.archived) liveGoalIdsNonPaginated.push(g.id);
-				}
-			}
-			// BFS: live parents/goals → their archived children → children of those, etc.
-			const archivedDelegatesOfLive = bfsEnrichArchived([...liveIdSet, ...liveGoalIdsNonPaginated], allArchivedForBfsNonPaginated);
-			json({ generation: currentGen, sessions, archivedDelegates: archivedDelegatesOfLive });
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/activate-skill — autonomous skill activation
-	const activateSkillMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/activate-skill$/);
-	if (activateSkillMatch && req.method === "POST") {
-		const sessionId = activateSkillMatch[1];
-		const session = sessionManager.getSession(sessionId);
-		if (!session) {
-			json({ error: "Session not found" }, 404);
-			return;
-		}
-		const body = await readBody(req);
-		const skillName = typeof body?.name === "string" ? body.name : "";
-		const skillArgs = typeof body?.args === "string" ? body.args : "";
-		if (!skillName) {
-			json({ error: "name is required" }, 400);
-			return;
-		}
-		// Resolve skill discovery context: host-side cwd + per-project store.
-		let resolvedConfigStore: { get(key: string): string | undefined } | undefined = projectConfigStore;
-		let skillCwd = session.cwd;
-		if (session.projectId) {
-			const pcm = (sessionManager as any).projectContextManager as import("./agent/project-context-manager.js").ProjectContextManager | undefined;
-			const ctx = pcm?.getOrCreate(session.projectId);
-			if (ctx) {
-				resolvedConfigStore = ctx.projectConfigStore;
-				if (session.sandboxed) skillCwd = ctx.project.rootPath;
-			}
-		}
-		const skill = getSlashSkill(skillCwd, skillName, resolvedConfigStore);
-		if (!skill) {
-			json({ error: `Skill "${skillName}" not found` }, 404);
-			return;
-		}
-		if (skill.disableModelInvocation === true) {
-			json({ error: `Skill "${skillName}" has disable-model-invocation: true and cannot be activated by the model` }, 403);
-			return;
-		}
-		// Inject the activation header so autonomous activation is byte-equal
-		// to user `/<name>` invocation.
-		const pathRewrite = session.sandboxed
-			? (hostPath: string): string | null => {
-				// Project worktree mounts at /workspace; rewrite when the host
-				// path lives under it. Built-in / personal skills aren't mounted.
-				const projectRoot = (session.projectId
-					? ((sessionManager as any).projectContextManager as import("./agent/project-context-manager.js").ProjectContextManager | undefined)?.getOrCreate(session.projectId)?.project.rootPath
-					: undefined);
-				const normHost = hostPath.replace(/\\/g, "/");
-				const normProj = projectRoot ? projectRoot.replace(/\\/g, "/") : null;
-				const sessionCwdNorm = session.cwd.replace(/\\/g, "/");
-				for (const candidate of [normProj, sessionCwdNorm]) {
-					if (candidate && (normHost === candidate || normHost.startsWith(candidate + "/"))) {
-						const rel = normHost.slice(candidate.length).replace(/^\/+/, "");
-						return "/workspace" + (rel ? "/" + rel : "");
-					}
-				}
-				return null;
-			}
-			: undefined;
-		const skillBody = buildSlashSkillPrompt(skill, skillArgs);
-		const expanded = buildActivationHeader(skill, pathRewrite) + skillBody;
-		json({ ok: true, expanded, source: skill.source, filePath: skill.filePath });
-		return;
-	}
-
-	// POST /api/sessions/:id/tool-grant-request — long-polling endpoint called by guard extension
-	const toolGrantMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/tool-grant-request$/);
-	if (toolGrantMatch && req.method === "POST") {
-
-		const sessionId = toolGrantMatch[1];
-		const body = await readBody(req);
-		if (!body || !body.toolName || !body.toolGroup) {
-			json({ error: "toolName and toolGroup required" }, 400);
-			return;
-		}
-		try {
-			const result = await sessionManager.requestToolGrant(sessionId, body.toolName, body.toolGroup);
-			json(result);
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// GET /api/sessions/:id (exact match — not /api/sessions/:id/output etc.)
-	const singleSessionMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)$/);
-	if (singleSessionMatch && req.method === "GET") {
-		const id = singleSessionMatch[1];
-		const session = sessionManager.getSession(id);
-		if (!session) {
-			// Check if it's an archived session
-			const archived = sessionManager.getArchivedSession(id);
-			if (archived) {
-				json({
-					id: archived.id,
-					title: archived.title,
-					cwd: archived.cwd,
-					projectId: archived.projectId,
-					status: "archived",
-					createdAt: archived.createdAt,
-					lastActivity: archived.lastActivity,
-					clientCount: 0,
-					isCompacting: false,
-					goalId: archived.goalId,
-					assistantType: archived.assistantType,
-					delegateOf: archived.delegateOf,
-					role: archived.role,
-					teamGoalId: archived.teamGoalId,
-					teamLeadSessionId: archived.teamLeadSessionId,
-					worktreePath: archived.worktreePath,
-					taskId: archived.taskId,
-					staffId: archived.staffId,
-					colorIndex: colorStore.get(archived.id),
-					preview: archived.preview,
-					reattemptGoalId: archived.reattemptGoalId,
-					archived: true,
-					archivedAt: archived.archivedAt,
-					imageGenerationModel: sessionManager.getImageModelForSession(archived.id),
-				});
-				return;
-			}
-			res.writeHead(404);
-			res.end(JSON.stringify({ error: "Session not found" }));
-			return;
-		}
-		const sessionPs = sessionManager.getSessionStore(session.projectId).get(session.id);
-		json({
-			id: session.id,
-			title: session.title,
-			cwd: session.cwd,
-			status: session.status,
-			createdAt: session.createdAt,
-			lastActivity: session.lastActivity,
-			clientCount: session.clients.size,
-			isCompacting: session.isCompacting,
-			goalId: session.goalId,
-			assistantType: session.assistantType,
-			// Legacy boolean fields for backward compat
-			goalAssistant: session.assistantType === "goal",
-			roleAssistant: session.assistantType === "role",
-			toolAssistant: session.assistantType === "tool",
-			delegateOf: session.delegateOf,
-			role: session.role,
-			teamGoalId: session.teamGoalId,
-			teamLeadSessionId: session.teamLeadSessionId,
-			worktreePath: session.worktreePath,
-			branch: session.branch ?? sessionPs?.branch,
-			taskId: session.taskId,
-			staffId: session.staffId,
-			colorIndex: colorStore.get(session.id),
-			preview: session.preview,
-			reattemptGoalId: sessionPs?.reattemptGoalId,
-			projectId: sessionPs?.projectId || session.projectId,
-			// Persisted model selection (provider+id). Surfaces the result of
-			// the WS `set_model` handler's `persistSessionModel` call so clients
-			// (and tests) can verify the selection round-tripped to disk without
-			// reaching into the WS state stream.
-			modelProvider: sessionPs?.modelProvider,
-			modelId: sessionPs?.modelId,
-			restoreError: session.restoreError,
-			lastTurnErrored: session.lastTurnErrored ?? false,
-			consecutiveErrorTurns: session.consecutiveErrorTurns ?? 0,
-			completedTurnCount: session.completedTurnCount ?? 0,
-			imageGenerationModel: sessionManager.getImageModelForSession(session.id),
-		});
-		return;
 	}
 
 	// POST /api/sessions
@@ -2456,98 +2137,6 @@ async function handleApiRoute(
 	// ── Team endpoints ─────────────────────────────────────────────
 	// Routes accept both /team/ and legacy /swarm/ paths
 
-	// Routes with :id parameter
-	const sessionMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)$/);
-	if (sessionMatch) {
-		const id = sessionMatch[1];
-
-		if (req.method === "GET") {
-			const session = sessionManager.getSession(id);
-			if (!session) {
-				json({ error: "Session not found" }, 404);
-				return;
-			}
-			json({
-				id: session.id,
-				title: session.title,
-				cwd: session.cwd,
-				status: session.status,
-				createdAt: session.createdAt,
-				clientCount: session.clients.size,
-			});
-			return;
-		}
-
-		if (req.method === "DELETE") {
-			const purge = url.searchParams.get("purge") === "true";
-			// Check if it's an archived session — purge immediately
-			const archivedSession = sessionManager.getArchivedSession(id);
-			if (archivedSession) {
-				await sessionManager.purgeArchivedSession(id);
-				json({ ok: true });
-				return;
-			}
-			const terminated = await sessionManager.terminateSession(id);
-			if (!terminated) {
-				// Session not in memory — check if it's a dormant store entry (e.g. completed delegate)
-				if (purge) {
-					// Archive it first so purge can find it, then purge
-					sessionManager.storeArchive(id);
-					const purged = await sessionManager.purgeArchivedSession(id);
-					if (purged) {
-						json({ ok: true });
-						return;
-					}
-				}
-				json({ error: "Session not found" }, 404);
-				return;
-			}
-			// If purge requested, also purge the now-archived session immediately
-			if (purge) {
-				await sessionManager.purgeArchivedSession(id);
-			}
-			json({ ok: true });
-			return;
-		}
-	}
-
-	// POST /api/sessions/:id/wait — block until session becomes idle
-	// Uses chunked transfer with periodic heartbeat newlines to prevent
-	// HTTP client body-timeout (undici defaults to 300s between chunks).
-	const waitMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/wait$/);
-	if (waitMatch && req.method === "POST") {
-		const id = waitMatch[1];
-		const body = await readBody(req);
-		const timeoutMs = body?.timeout_ms ?? 600_000;
-
-		// Stream chunked response with heartbeat to keep connection alive
-		res.writeHead(200, {
-			"Content-Type": "application/json",
-			"Transfer-Encoding": "chunked",
-			"Cache-Control": "no-cache",
-		});
-
-		// Send a heartbeat newline every 60s to prevent client body-timeout
-		const heartbeat = setInterval(() => {
-			try { res.write("\n"); } catch { /* connection gone */ }
-		}, 60_000);
-
-		try {
-			await sessionManager.waitForIdle(id, timeoutMs);
-			const output = await sessionManager.getSessionOutput(id);
-			const session = sessionManager.getSession(id);
-			res.end(JSON.stringify({
-				status: session?.status || "idle",
-				output,
-			}));
-		} catch (err) {
-			res.end(JSON.stringify({ error: String(err) }));
-		} finally {
-			clearInterval(heartbeat);
-		}
-		return;
-	}
-
 	// POST /api/sessions/:archivedId/continue — Continue-Archived (lossless)
 	//
 	// Clones the archived session's `.jsonl` into a fresh slot, registers it
@@ -2702,128 +2291,6 @@ async function handleApiRoute(
 			status: newSession.status,
 			title: continuedTitle,
 		}, 201);
-		return;
-	}
-
-	// PATCH /api/sessions/:id — update session properties (title, colorIndex, etc.)
-	const patchMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)$/);
-	if (patchMatch && req.method === "PATCH") {
-		const id = patchMatch[1];
-		const body = await readBody(req);
-		if (!body || typeof body !== "object") {
-			json({ error: "Invalid body" }, 400);
-			return;
-		}
-
-		if (typeof body.title === "string") {
-			const ok = sessionManager.setTitle(id, body.title);
-			if (!ok) { json({ error: "Session not found" }, 404); return; }
-		}
-
-		if (typeof body.colorIndex === "number") {
-			if (body.colorIndex < 0 || body.colorIndex > 13) {
-				json({ error: "colorIndex must be 0-13" }, 400);
-				return;
-			}
-			colorStore.set(id, body.colorIndex);
-		}
-
-		if (typeof body.projectId === "string") {
-			const session = sessionManager.getSession(id);
-			if (!session) { json({ error: "Session not found" }, 404); return; }
-			const oldProjectId = session.projectId;
-			const newProjectId = body.projectId || undefined;
-			session.projectId = newProjectId;
-			// Update in both old and new project stores to ensure consistency
-			sessionManager.getSessionStore(oldProjectId).update(id, { projectId: newProjectId });
-			if (newProjectId !== oldProjectId) {
-				sessionManager.getSessionStore(newProjectId).update(id, { projectId: newProjectId });
-			}
-		}
-
-		if (typeof body.preview === "boolean") {
-			const session = sessionManager.getSession(id);
-			if (!session) { json({ error: "Session not found" }, 404); return; }
-			session.preview = body.preview;
-			sessionManager.persistSessionMetadata(session).catch(() => {});
-			broadcastToAll({ type: "preview_changed", sessionId: id, preview: body.preview });
-		}
-
-		if (typeof body.roleId === "string" && body.roleId !== "") {
-			const role = roleManager.getRole(body.roleId);
-			if (!role) { json({ error: `Role "${body.roleId}" not found` }, 404); return; }
-			try {
-				const ok = await sessionManager.assignRole(id, role);
-				if (!ok) { json({ error: "Session not found" }, 404); return; }
-			} catch (err) {
-				jsonError(400, err);
-				return;
-			}
-		} else if (typeof body.roleId === "string" && body.roleId === "") {
-			// Clear role assignment
-			const session = sessionManager.getSession(id);
-			if (session) {
-				session.role = undefined;
-				session.accessory = undefined;
-				sessionManager.persistSessionMetadata(session).catch(() => {});
-			}
-		}
-
-		if (typeof body.assistantType === "string" || typeof body.goalAssistant === "boolean" || typeof body.goalId === "string") {
-			const session = sessionManager.getSession(id);
-			if (!session) { json({ error: "Session not found" }, 404); return; }
-			if (typeof body.assistantType === "string") session.assistantType = body.assistantType || undefined;
-			else if (typeof body.goalAssistant === "boolean") session.assistantType = body.goalAssistant ? "goal" : undefined;
-			if (typeof body.goalId === "string") session.goalId = body.goalId;
-			sessionManager.persistSessionMetadata(session).catch(() => {});
-		}
-
-		if (typeof body.accessory === "string") {
-			const session = sessionManager.getSession(id);
-			if (session) {
-				session.accessory = body.accessory || undefined;
-				sessionManager.persistSessionMetadata(session).catch(() => {});
-			}
-		}
-
-		if (typeof body.delegateOf === "string") {
-			const session = sessionManager.getSession(id);
-			if (session) {
-				session.delegateOf = body.delegateOf || undefined;
-				sessionManager.updateSessionMeta(id, { delegateOf: body.delegateOf || undefined });
-			} else {
-				sessionManager.updateSessionMeta(id, { delegateOf: body.delegateOf || undefined });
-			}
-		}
-
-		if (typeof body.teamLeadSessionId === "string") {
-			// Update teamLeadSessionId — works for both live and archived sessions
-			const session = sessionManager.getSession(id);
-			if (session) {
-				sessionManager.updateSessionMeta(id, { teamLeadSessionId: body.teamLeadSessionId });
-			} else {
-				// Try archived session — update store directly
-				const archived = sessionManager.getArchivedSession(id);
-				if (archived) {
-					sessionManager.updateArchivedMeta(id, { teamLeadSessionId: body.teamLeadSessionId });
-				} else {
-					json({ error: "Session not found" }, 404); return;
-				}
-			}
-		}
-
-		if (body.archived === true) {
-			// Try to terminate live session first (which archives it)
-			const session = sessionManager.getSession(id);
-			if (session) {
-				try { await sessionManager.terminateSession(id); } catch {}
-			} else {
-				// Dormant/store-only session — archive directly in the store
-				sessionManager.storeArchive(id);
-			}
-		}
-
-		json({ ok: true });
 		return;
 	}
